[ANDROSDK-1974] fix package names
