+++ core/src/main/java/org/hisp/dhis/android/core/trackedentity/TrackedEntityInstanceFilterCollectionRepository.kt
@@ -44,7 +44,7 @@ import org.hisp.dhis.android.core.trackedentity.internal.EntityQueryCriteriaFiel
 import org.hisp.dhis.android.core.trackedentity.internal.TrackedEntityInstanceFilterAttributeValueFilterChildrenAppender
 import org.hisp.dhis.android.core.trackedentity.internal.TrackedEntityInstanceFilterEvenFilterChildrenAppender
 import org.hisp.dhis.android.core.trackedentity.internal.TrackedEntityInstanceFilterStore
-import org.hisp.dhis.android.network.trackedEntityInstanceFilter.TrackedEntityInstanceFilterFields
+import org.hisp.dhis.android.network.trackedentityinstancefilter.TrackedEntityInstanceFilterFields
 import org.koin.core.annotation.Singleton
 
 @Singleton
