+++ core/src/main/java/org/hisp/dhis/android/network/eventFilter/EventFilterDTO.kt
@@ -26,7 +26,7 @@
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.network.eventFilter
+package org.hisp.dhis.android.network.eventfilter
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
