+++ core/src/main/java/org/hisp/dhis/android/network/eventFilter/EventFilterService.kt
@@ -25,7 +25,7 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.android.network.eventFilter
+package org.hisp.dhis.android.network.eventfilter
 
 import org.hisp.dhis.android.core.event.EventFilter
 import org.hisp.dhis.android.network.common.HttpServiceClientKotlinx
