+++ core/src/main/java/org/hisp/dhis/android/network/trackedentityinstancefilter/TrackedEntityInstanceFilterService.kt
@@ -25,21 +25,47 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+package org.hisp.dhis.android.network.trackedentityinstancefilter
 
-package org.hisp.dhis.android.network.trackedEntityInstanceFilter
+import org.hisp.dhis.android.core.trackedentity.TrackedEntityInstanceFilter
+import org.hisp.dhis.android.network.common.HttpServiceClientKotlinx
+import org.hisp.dhis.android.network.common.fields.Fields
+import org.hisp.dhis.android.network.common.filters.Filter
+import org.koin.core.annotation.Singleton
 
-import kotlinx.serialization.Serializable
-import org.hisp.dhis.android.core.common.FilterPeriod
+@Singleton
+internal class TrackedEntityInstanceFilterService(private val client: HttpServiceClientKotlinx) {
+    suspend fun getTrackedEntityInstanceFilters(
+        uids: Filter<TrackedEntityInstanceFilter>,
+        accessDataReadFilter: String,
+        fields: Fields<TrackedEntityInstanceFilter>,
+        paging: Boolean,
+    ): TrackedEntityInstanceFilterPayload {
+        return client.get {
+            url("trackedEntityInstanceFilters")
+            parameters {
+                fields(fields)
+                filter(uids)
+                attribute("filter", accessDataReadFilter)
+                paging(paging)
+            }
+        }
+    }
 
-@Serializable
-internal data class FilterPeriodDTO(
-    val periodFrom: Int?,
-    val periodTo: Int?,
-) {
-    fun toDomain(): FilterPeriod {
-        return FilterPeriod.builder()
-            .periodFrom(periodFrom)
-            .periodTo(periodTo)
-            .build()
+    suspend fun getTrackedEntityInstanceFilters37(
+        uids: Filter<TrackedEntityInstanceFilter>,
+        accessDataReadFilter: String,
+        fields: Fields<TrackedEntityInstanceFilter>,
+        paging: Boolean,
+    ): TrackedEntityInstanceFilter37Payload {
+        return client.get {
+            url("trackedEntityInstanceFilters")
+            parameters {
+                fields(fields)
+                filter(uids)
+                attribute("filter", accessDataReadFilter)
+                paging(paging)
+            }
+        }
     }
 }
