+++ core/src/test/java/org/hisp/dhis/android/core/arch/api/fields/internal/FieldsStringGenerationShould.kt
@@ -113,14 +113,14 @@ import org.hisp.dhis.android.network.categoryoption.CategoryOptionFields
 import org.hisp.dhis.android.network.common.fields.Fields
 import org.hisp.dhis.android.network.common.fields.Property
 import org.hisp.dhis.android.network.constant.ConstantFields
-import org.hisp.dhis.android.network.customIcon.CustomIconFields
-import org.hisp.dhis.android.network.eventFilter.EventFilterFields
+import org.hisp.dhis.android.network.customicon.CustomIconFields
+import org.hisp.dhis.android.network.eventfilter.EventFilterFields
 import org.hisp.dhis.android.network.legendset.LegendSetFields
 import org.hisp.dhis.android.network.option.OptionFields
 import org.hisp.dhis.android.network.optiongroup.OptionGroupFields
 import org.hisp.dhis.android.network.optionset.OptionSetFields
-import org.hisp.dhis.android.network.trackedEntityInstanceFilter.TrackedEntityInstanceFilterFields
 import org.hisp.dhis.android.network.trackedentityattribute.TrackedEntityAttributeFields
+import org.hisp.dhis.android.network.trackedentityinstancefilter.TrackedEntityInstanceFilterFields
 import org.hisp.dhis.android.network.trackedentitytype.TrackedEntityTypeFields
 import org.hisp.dhis.android.network.trackervisualization.TrackerVisualizationDimensionFields
 import org.hisp.dhis.android.network.trackervisualization.TrackerVisualizationDimensionRepetitionFields
