+++ core/src/test/java/org/hisp/dhis/android/core/event/EventFilterShould.kt
@@ -29,9 +29,13 @@ package org.hisp.dhis.android.core.event
 
 import com.google.common.truth.Truth.assertThat
 import org.hisp.dhis.android.core.arch.helpers.DateUtils
-import org.hisp.dhis.android.core.common.*
+import org.hisp.dhis.android.core.common.AssignedUserMode
+import org.hisp.dhis.android.core.common.BaseObjectKotlinxShould
+import org.hisp.dhis.android.core.common.DatePeriodType
+import org.hisp.dhis.android.core.common.ObjectShould
+import org.hisp.dhis.android.core.common.RelativePeriod
 import org.hisp.dhis.android.core.organisationunit.OrganisationUnitMode
-import org.hisp.dhis.android.network.eventFilter.EventFilterDTO
+import org.hisp.dhis.android.network.eventfilter.EventFilterDTO
 import org.junit.Test
 
 class EventFilterShould : BaseObjectKotlinxShould("event/event_filter.json"), ObjectShould {
